+++ src/modules/comment/comment.service.ts
@@ -11,7 +11,6 @@ import { PaginateResult, Types } from 'mongoose';
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@app/transformers/model.transformer';
 import { getGuestbookPageUrl, getArticleUrl } from '@app/transformers/urlmap.transformer';
-import { parseMarkdownToHtml } from '@app/transformers/markdown.transformer';
 import { MongooseModel } from '@app/interfaces/mongoose.interface';
 import { ECommentPostType, ECommentState } from '@app/interfaces/state.interface';
 import { IPService } from '@app/processors/helper/helper.service.ip';
@@ -37,18 +36,20 @@ export class CommentService {
 
   // 邮件通知网站主及目标对象
   private sendMailToAdminAndTargetUser(comment: Comment, permalink: string) {
-    const commentHtml = parseMarkdownToHtml(comment.content);
-    const commentType = comment.post_id === ECommentPostType.Guestbook ? '留言' : '评论';
-    const getContextPrefix = isReply => `来自 ${comment.author.name} 的${commentType}${isReply ? '回复' : ''}：`;
+    const commentTypeText = comment.post_id === ECommentPostType.Guestbook ? '留言' : '评论';
+    const getContextPrefix = isReply => {
+      const replyText = isReply ? '回复' : '';
+      return `来自 ${comment.author.name} 的${commentTypeText}${replyText}：`
+    };
     const sendMailText = contentPrefix => `${contentPrefix}${comment.content}`;
     const sendMailHtml = contentPrefix => `
-      <p>${contentPrefix}${commentHtml}</p><br>
+      <p>${contentPrefix}${comment.content}</p><br>
       <a href="${permalink}" target="_blank">[ 点击查看 ]</a>
     `;
 
     this.emailService.sendMail({
       to: APP_CONFIG.EMAIL.admin,
-      subject: `博客有新的${commentType}`,
+      subject: `博客有新的${commentTypeText}`,
       text: sendMailText(getContextPrefix(false)),
       html: sendMailHtml(getContextPrefix(false)),
     });
@@ -59,7 +60,7 @@ export class CommentService {
         .then(parentComment => {
           this.emailService.sendMail({
             to: parentComment.author.email,
-            subject: `你在 ${APP_CONFIG.APP.NAME} 有新的${commentType}回复`,
+            subject: `你在 ${APP_CONFIG.APP.NAME} 有新的${commentTypeText}回复`,
             text: sendMailText(getContextPrefix(true)),
             html: sendMailHtml(getContextPrefix(true)),
           });