+++ CHANGELOG.md
@@ -1,8 +1,45 @@
 
-## TODO
-- 删除一条存在子级评论的评论时，子级评论的 PID 需要置空
+# Changelog
+All notable changes to this project will be documented in this file.
 
-## CHANGELOG
+### v3.4.4
+
+**Feature**
+- DB backup shell & path
+- Support patch API to backup DB & recover DB
+
+### v3.4.3
+
+**Chore**
+- Upgrade all deps
+- Remove unused deps
+- Update API Document
+- Fix typos
+- Add `cspell.json`
+
+**Breaking Change**
+- Remove `Vlog` `Music` `GitHub` modules to BFF server
+
+### v3.4.2
+
+**Chore**
+- Upgrade all deps
+
+**Breaking Change**
+- comment content HTML -> markdown text
+- remove marked module
+
+### v3.4.1
+
+**Chore**
+- Upgrade all deps
+
+**Feature**
+- support redis `password` config
+
+**Breaking Change**
+- add fe public path
+- remove legacy folder file path for syndication
 
 ### v3.4.0
 - Upgrade deps